Domain basics was modified
