--- src/main/java/com/solvd/qaprotours/domain/CateringType.java
@@ -1,11 +1,0 @@
-package com.solvd.qaprotours.domain;
-
-/**
- * @author Ermakovich Kseniya
- */
-public enum CateringType {
-
-    BREAKFAST,
-    ALL_INCLUSIVE
-
-}
