--- src/main/java/com/solvd/qaprotours/domain/TourType.java
@@ -1,12 +1,0 @@
-package com.solvd.qaprotours.domain;
-
-/**
- * @author Ermakovich Kseniya
- */
-public enum TourType {
-
-    HEALTH,
-    CULTURE
-
-}
-
