+++ src/main/java/com/solvd/qaprotours/domain/tour/Tour.java
@@ -1,5 +1,6 @@
-package com.solvd.qaprotours.domain;
+package com.solvd.qaprotours.domain.tour;
 
+import com.solvd.qaprotours.domain.hotel.Hotel;
 import com.solvd.qaprotours.domain.user.Ticket;
 import jakarta.persistence.*;
 import lombok.Data;
@@ -23,13 +24,9 @@ public class Tour {
     private String name;
     private String description;
 
-    @ManyToOne
-    @JoinColumn(name = "country_id")
-    private Country country;
+    private String country;
 
-    @ManyToOne
-    @JoinColumn(name = "city_id")
-    private City city;
+    private String city;
 
     @Enumerated(EnumType.STRING)
     private TourType type;
@@ -55,5 +52,19 @@ public class Tour {
 
     @OneToMany(mappedBy = "tour", orphanRemoval = true)
     private List<Ticket> tickets;
+
+    public enum CateringType {
+
+        BREAKFAST,
+        ALL_INCLUSIVE
+
+    }
+
+    public enum TourType {
+
+        HEALTH,
+        CULTURE
+
+    }
 
 }
