--- src/main/java/com/solvd/qaprotours/domain/user/Role.java
@@ -1,21 +1,0 @@
-package com.solvd.qaprotours.domain.user;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
-
-/**
- * @author Ermakovich Kseniya
- */
-@RequiredArgsConstructor
-public enum Role implements GrantedAuthority {
-
-    CLIENT("CLIENT"),
-    EMPLOYEE("EMPLOYEE");
-
-    private final String value;
-
-    @Override
-    public String getAuthority() {
-        return "ROLE_" + value;
-    }
-}
