--- src/main/java/com/solvd/qaprotours/domain/user/Status.java
@@ -1,12 +1,0 @@
-package com.solvd.qaprotours.domain.user;
-
-/**
- * @author Ermakovich Kseniya
- */
-public enum Status {
-
-    ORDERED,
-    CONFIRMED
-
-}
-
