+++ src/main/java/com/solvd/qaprotours/domain/user/User.java
@@ -2,6 +2,8 @@ package com.solvd.qaprotours.domain.user;
 
 import jakarta.persistence.*;
 import lombok.Data;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.GrantedAuthority;
 
 /**
  * @author Ermakovich Kseniya
@@ -25,5 +27,19 @@ public class User {
 
     @Column(name = "is_activated")
     private boolean isActivated;
+
+    @RequiredArgsConstructor
+    public enum Role implements GrantedAuthority {
+
+        CLIENT("CLIENT"),
+        EMPLOYEE("EMPLOYEE");
+
+        private final String value;
+
+        @Override
+        public String getAuthority() {
+            return "ROLE_" + value;
+        }
+    }
 
 }
