+++ src/main/java/com/solvd/qaprotours/web/dto/AddressDto.java
@@ -1,7 +1,5 @@
 package com.solvd.qaprotours.web.dto;
 
-import com.solvd.qaprotours.domain.City;
-import com.solvd.qaprotours.domain.Country;
 import com.solvd.qaprotours.web.dto.validation.OnCreate;
 import com.solvd.qaprotours.web.dto.validation.OnUpdate;
 import jakarta.validation.constraints.NotNull;
@@ -18,10 +16,12 @@ public class AddressDto {
     private Long id;
 
     @NotNull(message = "country cannot be null", groups = {OnCreate.class, OnUpdate.class})
-    private Country country;
+    @Size(min=3, max = 40, message = "Country should be from 3 to 40 symbols")
+    private String country;
 
     @NotNull(message = "city cannot be null", groups = {OnCreate.class, OnUpdate.class})
-    private City city;
+    @Size(min=3, max = 40, message = "City should be from 3 to 40 symbols")
+    private String city;
 
     @NotNull(message = "street name cannot be null", groups = {OnCreate.class, OnUpdate.class})
     @Size(min = 1, max = 120, message = "street name must be between 1 and 120 characters", groups = {OnCreate.class, OnUpdate.class})
