+++ src/main/java/com/solvd/qaprotours/web/dto/TourDto.java
@@ -1,9 +1,6 @@
 package com.solvd.qaprotours.web.dto;
 
-import com.solvd.qaprotours.domain.CateringType;
-import com.solvd.qaprotours.domain.City;
-import com.solvd.qaprotours.domain.Country;
-import com.solvd.qaprotours.domain.TourType;
+import com.solvd.qaprotours.domain.tour.Tour;
 import com.solvd.qaprotours.web.dto.validation.OnCreate;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
@@ -31,16 +28,18 @@ public class TourDto {
     private String description;
 
     @NotNull(message = "country cannot be null", groups = {OnCreate.class})
-    private Country country;
+    @Size(min=3, max = 40, message = "Country should be from 3 to 40 symbols")
+    private String country;
 
     @NotNull(message = "city cannot be null", groups = {OnCreate.class})
-    private City city;
+    @Size(min=3, max = 40, message = "City should be from 3 to 40 symbols")
+    private String city;
 
     @NotNull(message = "type cannot be null", groups = {OnCreate.class})
-    private TourType type;
+    private Tour.TourType type;
 
     @NotNull(message = "catering type cannot be null", groups = {OnCreate.class})
-    private CateringType cateringType;
+    private Tour.CateringType cateringType;
 
     @NotNull(message = "hotel cannot be null", groups = {OnCreate.class})
     private HotelDto hotel;
