+++ src/main/java/com/solvd/qaprotours/web/dto/user/TicketDto.java
@@ -1,7 +1,7 @@
 package com.solvd.qaprotours.web.dto.user;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.solvd.qaprotours.domain.user.Status;
+import com.solvd.qaprotours.domain.user.Ticket;
 import com.solvd.qaprotours.web.dto.TourDto;
 import com.solvd.qaprotours.web.dto.validation.OnCreate;
 import com.solvd.qaprotours.web.dto.validation.OnUpdate;
@@ -30,7 +30,7 @@ public class TicketDto {
     private LocalDateTime orderTime;
 
     @JsonProperty(access = JsonProperty.Access.READ_ONLY)
-    private Status status;
+    private Ticket.Status status;
 
     @NotNull(message = "clients amount must not be null", groups = {OnCreate.class})
     @Size(min = 1, max = 30, message = "clients amount must be between 1 and 30", groups = {OnCreate.class})
