+++ src/main/java/com/solvd/qaprotours/web/mapper/TourMapper.java
@@ -1,6 +1,6 @@
 package com.solvd.qaprotours.web.mapper;
 
-import com.solvd.qaprotours.domain.Tour;
+import com.solvd.qaprotours.domain.tour.Tour;
 import com.solvd.qaprotours.web.dto.TourDto;
 import org.mapstruct.Mapper;
